Add "Preview" tab to the right sidebar
The preview tab makes possible to select the audio and video devices to
be used in a call, as well as showing a preview of them (so the user now
can check what is visible in the camera before joining a call).

In order to show the preview of the devices a media stream has to be
requested, which will ask the user for permissions (unless they have
been granted permanently). To prevent permissions requests as soon as
the conversation is opened the preview tab is not shown as the first
tab, and to limit the use of the devices only to the scrictly necessary
the previews are enabled only when the preview tab is the active one.

Currently changing the devices during a call has no effect, so the
device selectors (as well as the previews themselves, as they will have
the same content already shown in the call view) are disabled during
calls. The tab itself is not removed due to ordering issues when
removing and adding it back.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>